code: double

+++ main.cpp
@@ -2,7 +2,7 @@
 
 int main() {
 	std::cout << "Enter A and B: ";
-	int a, b;
+	double a, b;
 	std::cin >> a >> b;
 	std::cout << "A + B = " << a + b << '\n'
 			  << "A - B = " << a - b << '\n'
